Add 'visible' method on `ActivityController` for making the Activity visible
This method adds the Activity's window decor view to the window manager. This is required
for `View.isShown` to return true as otherwise the Activity's view hierarchy won't have a parent.

In real life this happens some time after the Activity resumes (in `ActivityThread`) so it really needs
to be its own lifecycle operation (even though its not part of the 'official' lifecycle).